requirements.txt
annotated-types==0.7.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
docker-pycreds==0.4.0
eval_type_backport==0.2.2
filelock==3.16.1
fonttools==4.57.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
importlib_resources==6.4.5
Jinja2==3.1.6
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
packaging==24.2
pillow==10.4.0
platformdirs==4.3.6
protobuf==5.29.4
psutil==7.0.0
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
sentry-sdk==2.25.1
setproctitle==1.3.5
six==1.17.0
smmap==5.0.2
sympy==1.13.3
torch==2.4.1
typing_extensions==4.13.2
urllib3==2.2.3
wandb==0.19.9
zipp==3.20.2
